Do warehouse task

+++ api/urls/urls.controllers.js
@@ -1,23 +1,24 @@
-const Url = require('../../models/Url');
-const shortid = require('shortid');
-const User = require('../../models/User');
+const Url = require("../../models/Url");
+const shortid = require("shortid");
+const User = require("../../models/User");
 
-const baseUrl = 'http:localhost:8000';
+const baseUrl = "http:localhost:8000";
 
 exports.shorten = async (req, res) => {
   // create url code
   const urlCode = shortid.generate();
   try {
-    req.body.shortUrl = baseUrl + '/' + urlCode;
+    req.body.shortUrl = baseUrl + "/" + urlCode;
     req.body.urlCode = urlCode;
-    req.body.userId = req.params.userId;
+    req.body.userId = req.user._id;
+    console.log(req.body);
     const newUrl = await Url.create(req.body);
-    await User.findByIdAndUpdate(req.params.userId, {
+    await User.findByIdAndUpdate(req.user._id, {
       $push: { urls: newUrl._id },
     });
     res.json(newUrl);
   } catch (err) {
-    res.status(500).json('Server Error');
+    res.status(500).json("Server Error");
   }
 };
 
@@ -27,10 +28,10 @@ exports.redirect = async (req, res) => {
     if (url) {
       return res.redirect(url.longUrl);
     } else {
-      return res.status(404).json('No URL Found');
+      return res.status(404).json("No URL Found");
     }
   } catch (err) {
-    res.status(500).json('Server Error');
+    res.status(500).json("Server Error");
   }
 };
 
@@ -38,12 +39,15 @@ exports.deleteUrl = async (req, res) => {
   try {
     const url = await Url.findOne({ urlCode: req.params.code });
     if (url) {
+      if (url.userId.toString() !== req.user._id.toString())
+        return res.status(401).json("Not authorized to delete");
+
       await Url.findByIdAndDelete(url._id);
-      return res.status(201).json('Deleted');
+      return res.status(201).json("Deleted");
     } else {
-      return res.status(404).json('No URL Found');
+      return res.status(404).json("No URL Found");
     }
   } catch (err) {
-    res.status(500).json('Server Error');
+    res.status(500).json("Server Error");
   }
 };
